+++ Editor/Pooling/ObjectPoolMenu.cs
@@ -8,14 +8,14 @@ namespace Codetox.Editor.GameEvents
     public static class ObjectPoolMenu
     {
         [MenuItem("GameObject/" + Framework.MenuRoot.Pooling.GameObject)]
-        [MenuItem(Framework.MenuRoot.Pooling.GameObject)]
+        [MenuItem("Tools/" + Framework.MenuRoot.Pooling.GameObject)]
         public static void CreateGameObjectPool(MenuCommand command)
         {
             EditorHelper.CreateObjectOfType<GameObjectPool>(command.context as GameObject);
         }
 
         [MenuItem("GameObject/" + Framework.MenuRoot.Pooling.ParticleSystem)]
-        [MenuItem(Framework.MenuRoot.Pooling.ParticleSystem)]
+        [MenuItem("Tools/" + Framework.MenuRoot.Pooling.ParticleSystem)]
         public static void CreateParticleSystemPool(MenuCommand command)
         {
             EditorHelper.CreateObjectOfType<ParticleSystemPool>(command.context as GameObject);